kafka_producer/requirements.txt
kafka-python
python-dotenv
logging
prometheus-client
redis